Fix licenses for src/qttableview.*
One can not rip out parts of licensed software, mention the copyright holder
and say: It was $foo, copyright holder was $bar. But i ripped this out, you
are free to do with this what you want and without limitation - this will not
work that way.

So the copied parts remain under both QPL-1.0 and GPL2. The added code inherit
the original licenses as not stated otherwise.