Removed the cause of a huge memory leak
This patch removes a huge memory leak in `Proc::refresh()`.

NOTE: It may seem that there is still a relatively smaller memory leak in setting the graph as the tray and window icon. However, the increase in memory usage stops after a while. So, this closes https://github.com/lxqt/qps/issues/101