More fixes for the header, especially for its DND
(1) The command field is always the last section if existing. It can't be dragged.

(2) When the tree is enabled, the PID field is always the first section and can't be dragged. If nonexistent, it'll be created once the tree is enabled. If PID is removed, the tree will be removed too.

(3) Field addition by right clicking is fixed (previously, the position of the new field was incorrect).

NOTE: The header DND isn't very predictable in qps and this patch doesn't include a fix for that.

Closes https://github.com/lxqt/qps/issues/164