Made column DND predictable — and much more
Closes https://github.com/lxqt/qps/issues/166

Also:

(1) In most process monitors, the downward sort indicator has different meanings for different fields. For example, it means sorting from high to low for CPU and memory, while it shows the alphabetical order for strings (a → z). Qps seems to follow the same rule. Therefore, the default indicator is changed from "SortUp" to "SortDown".

(2) A small mistake is fixed in adding columns. Previously, the PID field would be loaded as the first column with the next startup even if it was moved by DND and there was no tree).

(3) A bug is fixed in column removal. Previously, if the removed column was positioned before the sorted one, sorting would become incorrect until the next session.

NOTE: I think most GUI problems are fixed now.