Silenced compiler warnings about QtTableView
Also, "solved" the RTL problem by enforcing an LTR layout on the table and its header. This makes Qps usable with RTL languages for now. A real fix requires deep changes to the code, for which I might have time only later (I added "FIXME" comments to the code as a reminder).

Closes https://github.com/lxqt/qps/issues/190  and closes https://github.com/lxqt/qps/issues/191