Removed all codes related to the nonexistent "command menu"
Lots of unused, old-fashioned codes have been removed. If, at some point, we need to execute commands, we could add clean and up-to-date codes, but I don't think Qps' job is command execution.

The "Command" menu has been renamed to "Session". For now, it only includes the Quit item (as before) — other items are in their proper places.

Also,

 * Some unused flags are removed;
 * Some code blocks are simplified; and
 * Double saving of settings is prevented on quitting.

NOTE: Although this was a big step, the code still needs a great amount of cleanup.