Used an SVG icon for the app (#520)
Also, the icon shown in the tray for a moment before the graph is loaded is changed to the theme icon; previously, it was the hard-coded XPM icon.

Of course, if the active icon set has a separate `qps` icon, it'll be used instead of the new SVG icon, as is the case with any app.

Closes https://github.com/lxqt/qps/issues/519